rag-streamlit/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
torch>=2.0.0
langchain>=0.1.0
langchain-anthropic>=0.0.1
langchain-openai>=0.0.1
python-dotenv>=1.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.0
sudachipy>=0.6.8
sudachidict-core>=20230927
streamlit>=1.30.0
streamlit_elements==0.1.0